Extend tests to ChaseLev and MegaDeque.
However, there are significant problems right now.  I've been trying
to fix a divergence observed in MichaelScott.  But now I've not only
run into the same divergance with ChaseLev, but I also see invariant
breakage that will segfault the process if the unsafe versions of
vector operations are used.  (I disabled them for the moment.)