Remove one level of indirection for boxed ChaseLev
Previously, we had an `IORef` holding a mutable vector. We can
remove one level of indirection by using an `UnliftedArray` (i.e.,
an `ArrayArray#`) to hold the array instead.

Before:

     ChaseLevDeque -> MutVar# -> MVector -> Array#

Now:

     ChaseLevDeque -> ArrayArray# -> Array#

Note: we previously got the vector size from the `MVector`; now
we get it from the `Array#`. In many applications `MVector`s unbox,
so getting sizes from them is faster than getting them from `Array#`s.
However, this is *not* such an application: we're continually storing
and retrieving the vectors.